deleted utils.js and style.css
