+++ src/components/service/MosesService.js
@@ -17,9 +17,9 @@ export default class MosesService extends React.Component {
     this.handleSubmit = this.handleSubmit.bind(this);
   }
   componentWillReceiveProps(nextProps) {
-        if(this.isComplete !== nextProps.isComplete) {
-            this.parseProps(nextProps);
-        }
+    if (this.isComplete !== nextProps.isComplete) {
+      this.parseProps(nextProps);
+    }
   }
   parseProps(nextProps) {
     this.isComplete = nextProps.isComplete;
