--- src/components/service/annotation/style.css
@@ -1,4 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Roboto');
-body {
-  font-family: 'roboto';
-}
